requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
idna==3.3
multidict==5.2.0
psycopg2==2.9.1
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0

numpy~=1.21.3
opencv-python~=4.5.4.58
pyzbar~=0.1.8
Pillow~=8.4.0